Делит код из main.js на модули util.js и ads.js
